requirements.txt
flask==3.1.1
opencv-python-headless==4.12.0.88
numpy==2.3.1
matplotlib==3.10.3
werkzeug==3.1.3
Flask-Bootstrap~=3.3.7.1
WTForms~=3.2.1
pillow~=11.3.0